store: moved sql to standalone functions
This allows the same query to be used in different DB methods.
